INTEGRATE: stats catalog changes from open to lu
[git-p4: depot-paths = "//open/lu/dev/": change = 6383]

+++ farrago/src/net/sf/farrago/catalog/FarragoCatalogUtil.java
@@ -940,7 +940,7 @@ public abstract class FarragoCatalogUtil
     {
         index.setAnalyzeTime(createTimestamp());
         // FIXME: once catalog is integrated, use long value
-        index.setPageCount(pageCount.intValue());
+        index.setPageCount(pageCount);
     }
     
     public static void updateHistogram(
